\name{wlocpoly-package}
\alias{wlocpoly-package}
\alias{wlocpoly}
\docType{package}
\title{
\packageTitle{wlocpoly}
}
\description{
\packageDescription{wlocpoly}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wlocpoly}
\packageIndices{wlocpoly}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{wlocpoly}

Maintainer: \packageMaintainer{wlocpoly}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
